Recovered from Eclipse.
git-svn-id: https://subversion.assembla.com/svn/cmosher@6783 884d43a4-ffbe-47e9-90ca-fd8c37d7eb32

+++ OneFileCopy/src/OneFileCopyServer.java
@@ -20,7 +20,7 @@ public class OneFileCopy
 		int cBuf = cBufMax;
 		if (f.length() < cBuf)
 		{
-			cBuf = f.length();
+			cBuf = (int)f.length();
 		}
 
 		ServerSocket srv = new ServerSocket(60013);
